Some more work on splitting RPC calls into multiple files
