+++ rpc/src/main/scala/org/bitcoins/rpc/client/MessageCalls.scala
@@ -1,0 +1,36 @@
+package org.bitcoins.rpc.client
+
+import org.bitcoins.core.crypto.ECPrivateKey
+import org.bitcoins.core.protocol.P2PKHAddress
+import play.api.libs.json.JsString
+
+import scala.concurrent.Future
+
+/**
+  * RPC calls related to the message signing functionality
+  * in bitcoind
+  */
+protected trait MessageCalls extends Client with BitcoindCall{
+  def signMessage(address: P2PKHAddress, message: String): Future[String] = {
+    bitcoindCall[String](
+      "signmessage",
+      List(JsString(address.value), JsString(message)))
+  }
+
+  def signMessageWithPrivKey(
+                              key: ECPrivateKey,
+                              message: String): Future[String] = {
+    bitcoindCall[String](
+      "signmessagewithprivkey",
+      List(JsString(key.toWIF(network)), JsString(message)))
+  }
+
+  def verifyMessage(
+                     address: P2PKHAddress,
+                     signature: String,
+                     message: String): Future[Boolean] = {
+    bitcoindCall[Boolean](
+      "verifymessage",
+      List(JsString(address.value), JsString(signature), JsString(message)))
+  }
+}
