+++ rpc/src/main/scala/org/bitcoins/rpc/client/P2PCalls.scala
@@ -2,13 +2,15 @@ package org.bitcoins.rpc.client
 
 
 import org.bitcoins.rpc.jsonmodels.{GetNetworkInfoResult, Node, NodeBan}
+import org.bitcoins.rpc.serializers.JsonSerializers._
+import org.bitcoins.rpc.serializers.JsonReaders._
 import java.net.URI
 
-import play.api.libs.json.{JsBoolean, JsString}
+import play.api.libs.json.{JsBoolean, JsNumber, JsString}
 
 import scala.concurrent.Future
 
-trait P2PCalls extends Client{
+protected trait P2PCalls extends Client with BitcoindCall {
 
   def addNode(address: URI, command: String): Future[Unit] = {
     bitcoindCall[Unit](
@@ -53,6 +55,20 @@ trait P2PCalls extends Client{
 
   def ping(): Future[Unit] = {
     bitcoindCall[Unit]("ping")
+  }
+
+  def setBan(
+              address: URI,
+              command: String,
+              banTime: Int = 86400,
+              absolute: Boolean = false): Future[Unit] = {
+    bitcoindCall[Unit](
+      "setban",
+      List(
+        JsString(address.getAuthority),
+        JsString(command),
+        JsNumber(banTime),
+        JsBoolean(absolute)))
   }
 
   def setNetworkActive(activate: Boolean): Future[Unit] = {
